refactor : custom exception 경로 수정
