+++ src/main/java/com/seo92js/news_alpha_backend/common/handler/GlobalExceptionHandler.java
@@ -1,9 +1,9 @@
 package com.seo92js.news_alpha_backend.common.handler;
 
 import com.seo92js.news_alpha_backend.dto.ErrorResponse;
-import com.seo92js.news_alpha_backend.domain.ai.exception.DuplicateEmailException;
+import com.seo92js.news_alpha_backend.domain.member.exception.DuplicateEmailException;
 import com.seo92js.news_alpha_backend.common.exception.ErrorCode;
-import com.seo92js.news_alpha_backend.domain.ai.exception.MemberNotFoundException;
+import com.seo92js.news_alpha_backend.domain.member.exception.MemberNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.converter.HttpMessageNotReadableException;
