+++ src/main/java/com/seo92js/news_alpha_backend/domain/member/exception/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package com.seo92js.news_alpha_backend.domain.ai.exception;
+package com.seo92js.news_alpha_backend.domain.member.exception;
 
 import com.seo92js.news_alpha_backend.common.exception.BusinessException;
 import com.seo92js.news_alpha_backend.common.exception.ErrorCode;
